OBJ BMP loaders - Simple Mesh and lighting for GL
I have implemented simple BMP file loader and OBJ file loader that make
use of my sysFileLoader methods makinf it very easy to implemnt.
I have also made a simple mesh class and shader class to make it able to
have multiple instances of those things. I still have to make some changes
or additions to the mesh in order to have a easier way for adding the
vertex information.

Also, I have made a simple camera class and a Transformation class that
can be attached to objects or models to make them move rotate and scale.
(only Transform class). I have to add that functionality to the Camera
class as well.

-George Wulfers