dding compressed files that are needed!!
If you try to run mkae sure to unzip these three files.

bin.rar, lib.rar, Models.rar (make sure that there is only one directory
called models not two.

There is a dll that is needed to run the OpenGL version which located
inside of the bin folder. glew32.lib I think it is called.

-George Wulfers